Laravel 5.8, user id uses big int. Audit macro utilizes big int

+++ src/AppframeServiceProvider.php
@@ -68,8 +68,8 @@ class AppframeServiceProvider extends ServiceProvider
 
     private function registerBlueprintMacro(){
         Blueprint::macro('audit',function(){
-            $this->unsignedInteger('created_by')->nullable();
-            $this->unsignedInteger('updated_by')->nullable();
+            $this->unsignedBigInteger('created_by')->nullable();
+            $this->unsignedBigInteger('updated_by')->nullable();
             $this->timestamps();
             $this->foreign('created_by')->references('id')->on('users')->onUpdate('cascade')->onDelete('set null');
             $this->foreign('updated_by')->references('id')->on('users')->onUpdate('cascade')->onDelete('set null');
